extract active filters bar
