+++ styles/components/Search.module.scss
@@ -1,13 +1,5 @@
 @import "../utils/mixins";
 
-.filterTags {
-  margin-bottom: var(--nypl-space-xs);
-
-  button[id^="ts-filter-clear-all"] {
-    color: var(--nypl-colors-ui-link-primary);
-    text-decoration: underline;
-  }
-}
 .searchContainer {
   background-color: var(--nypl-colors-ui-bg-default);
   padding: var(--nypl-space-s) 0;
@@ -30,16 +22,6 @@
   gap: var(--nypl-space-xxs);
   svg {
     margin-top: var(--nypl-space-xxxs);
-  }
-}
-
-.auxSearchContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-top: var(--nypl-space-xs);
-  @media screen and (min-width: 600px) {
-    flex-direction: row-reverse;
   }
 }
